GB code fixed to work as intended
I was trying to write to tile ram while the display was on. Now I wait for vertical blank, disable the LCD while writing to the video ram.

Also I converted all tabs to spaces, mostly to make it a less annoying task to write the blog post. Editing text in a browser where you want to use tabs is really vexing.